Rework how sponsored talks are indicated
Originally this was just shown as part of the title, in parenthesis.
This change creates a data field for the sponsor.  It also removes the
sponsor parenthetical from headers and buttons, where it looked out of
place.  Instead the value of the data field is displayed as a subtitle
on the talk page, in a similar style as the speaker designations.